# Use more ideomatic Mojo and unify integer width with C. 