@martinvuyk 2024-10-19
It is actually correct that C int is `Int32` and Mojo's Int is the bitwidth of the platform (Int32 for 32-bit Int64 for 64-bit), but I don't think it will have any real performance impact in this comparison. And the comparison is (insofar as I've read the other code) more about "default behavior for a simple implementation" than overly optimized super unsafe benchmark.

In this case to be fair with the other code comparisons and because SIMD would mean changing your algorithm. You could also just go for `InlineArray[Int, 10]` which will have (Work In Progress) the same API as List but allocates on the stack and if you do `cache[digit]` with a compile time known for loop it will do the bounds check at compile time so all overhead magically disappears (List is on the heap and does bounds checks at runtime). You could also use `cache.unsafe_get(digit)` since there is no way that the while loop goes out of bounds, but to be fair with the other code (I'm sure C++ has some of these optimizations as well) let's leave it at default

Also, now `var` is optional (if you're not declaring the type explicitly (e.g. `var a: Int = 0`), otherwise it is still needed (in the future it shouldn't) )

Any op with numbers assumes `Int` by default like in Python

So overall this is pretty "idiomatic":
```mojo
alias N = 440_000_000


fn is_munchausen(number: Int, cache: InlineArray[Int, 10]) -> Bool:
    n = number
    total = 0

    while n > 0:
        digit = n % 10
        total += cache[digit]
        if total > number:
            return False
        n //= 10

    return total == number


fn get_cache() -> InlineArray[Int, 10]:
    ca = InlineArray[Int, 10]()
    ca[0] = 0

    @parameter
    for i in range(1, 10):
        ca[i] = i**i
    return ca


fn main():
    cache = get_cache()
    for n in range(0, N):
        if is_munchausen(n, cache):
            print(n)
```