@dorjeduck 2024-10-19
Great to have you both here, [owenhilyard](https://github.com/owenhilyard) and [martinvuyk](https://github.com/martinvuyk)! 
I agree with Martin's argument for keeping ```Int``` and against ```SIMD```, but I'll leave the final decision to you experts. When I tested the ```InlineArray``` based implementation, it ran slower than the current implementation on my local setup, both with Mojo 24.5 and the current nightly build.
Given the variability of nightly builds, I'd recommend avoiding reliance on them for SpeedTests and wait for the next stable release. That said, any improvements in performance is of course fantastic ;-)