@martinvuyk 2024-10-19
Hi @dorjeduck  that is weird, now that I read the code for the empty constructor for `InlineArray` it should actually fail at comp time (I had no LSP on lol). 

You made me curious and did some tests in my machine running only once (yeah not the best benchmark) `time magic run mojo run ./asd.mojo`:
- `InlineArray[Int, 10](0)`: `real    0m12,404s` # using index to set the items, has bounds checks
- `List[Int](capacity=10)`: `real    0m5,113s` # appending, has bounds checks
- `stack_allocation[10, Int]()`: `real    0m2,974s` # not printing correct numbers for some reason, so doesn't count
- `UnsafePointer[Int].alloc(10)`: `real    0m4,998s` # has no bounds checks, ofc adding a cache.free() at the end of main
- `SIMD[DType.index, 16]()`: `real    0m4,868s` # has no bounds checks

Come to think of it, the pointer is so small that it makes no difference in this case whether you allocate it on the heap or stack because you aren't doing anything that flushes the cache so it stays there very close and gets hit every time.

Now the even weirder parts:
- doing cache a compile time alias `alias cache = get_cache()` makes things even worse, in the case of `List[Int]`: `real    0m17,016s` when making it an alias in `main()` scope, takes more than 30s when declared as alias inside `is_munchausen()` scope

So overall it seems like `List[Int]` is a very solid default :)

I went on a tangent and took liberties with the code to try to make it run faster on SIMD, but I get `real    0m5,384s` :cry: 
```mojo
from sys.intrinsics import unlikely

alias N = 440_000_000


fn is_munchausen(number: Int) -> Bool:
    alias cache = get_cache()
    vec = SIMD[DType.uint32, 16]()

    @parameter
    for i in range(10):
        vec[i] = cache[(number // (10**i)) % 10]

    return int(vec.reduce_add()) == number


fn get_cache() -> SIMD[DType.uint32, 16]:
    ca = SIMD[DType.uint32, 16](0)

    @parameter
    for i in range(1, 10):
        ca[i] = i**i
    return ca


fn main():
    for n in range(0, N):
        if unlikely(is_munchausen(n)):
            print(n)
```

Anyway, this was fun :smile:  . @jabbalaci Nice repo and problem to get familiar with languages :muscle: