@dorjeduck 2024-10-20
thanks @martinvuyk for sharing your insights - love your creative SIMD based implementation