@jabbalaci 2024-10-20
Thank you guys! Should we add an optimized version too or can we leave the current version? I suggest leaving a "normal, starightforward" version. But next to `main.mojo`, a `main2.mojo` can be added at any time. See for instance the `nim2/` folder that contains several versions.