@dorjeduck 2024-10-20
Sounds great to me. ```List``` seems to work well for a "straightforward" version, with little cleanup of the code i previously submitted, something like

```python
alias N = 440_000_000

fn is_munchausen(number: Int, cache: List[Int]) -> Bool:
    var n = number
    var total:Int = 0

    while n > 0:
        var digit:Int = n % 10
        total += cache[digit]
        if total > number:
            return False
        n //= 10

    return total == number

fn get_cache() -> List[Int]:
    var ca = List[Int](capacity=10)
    ca.append(0)

    @parameter
    for i in range(1,10):
        ca.append(i**i)
    return ca

fn main():
    var cache = get_cache()
    for n in range(0, N):
        if is_munchausen(n, cache):
            print(n)
```
I believe there's great merit in having a version that's Python-like while delivering excellent performance. Equally interesting is having an optimized version alongside it, showcasing the more C-like aspects of Mojo.

For the optimized version, I step back and leave it to the experts ;-)

ps
i can make a PR for this slightly changed version, if there are no objections to this "normal" version. Or you just replace the current one with this one.