@jabbalaci 2024-10-20
A PR would be appreciated. But let's keep a straightforward version too.