@owenhilyard 2024-10-20
@jabbalaci It is closed because the CPU you are benchmarking on doesn't play nicely with this style of programming. It has vector registers which are half the size of what I benchmarked on, which is smaller than the cache is, so the compiler does a bunch of workarounds but ends up being much slower, since that generation also has frequency penalties for using 256 bit SIMD. On the Zen 5 CPU I tested it on, the version I posted was close to the vectorized C.