@diocletiann 2024-10-21
A simple change to `Uint32` took a full second off on my M1 Max.

```
alias N = 440_000_000

fn is_munchausen(number: UInt32, cache: List[UInt32]) -> Bool:
    n = number
    var total: UInt32 = 0

    while n > 0:
        digit = n % 10
        total += cache[int(digit)]
        if total > number:
            return False
        n //= 10

    return total == number

fn get_cache() -> List[UInt32]:
    ca = List[UInt32](capacity=10)
    ca.append(0)

    @parameter
    for i in range(1,10):
        ca.append(i**i)
    return ca

fn main():
    cache = get_cache()
    for n in range(0, N):
        if is_munchausen(n, cache):
            print(n)
```
Before:
```
Benchmark 1: ./munch
  Time (mean ± σ):      2.751 s ±  0.000 s    [User: 2.739 s, System: 0.012 s]
  Range (min … max):    2.751 s …  2.751 s    2 runs
```
After:
```
Benchmark 1: ./munch
  Time (mean ± σ):      1.726 s ±  0.016 s    [User: 1.710 s, System: 0.009 s]
  Range (min … max):    1.715 s …  1.737 s    2 runs
```